ai-animal-classification/train.py
import os
import torch
import torch.nn as nn
# import torchvision
from torchvision import datasets, transforms, models
from torch.utils.data import DataLoader
from tqdm import tqdm

# Paths
DATASET_DIR = "dataset"
MODEL_DIR = "model"
os.makedirs(MODEL_DIR, exist_ok=True)

# Config
IMG_SIZE = 180
BATCH_SIZE = 16
EPOCHS = 10
DEVICE = torch.device("cuda" if torch.cuda.is_available() else "cpu")

# 📦 Data transforms
transform = transforms.Compose([
    transforms.Resize((IMG_SIZE, IMG_SIZE)),
    transforms.RandomHorizontalFlip(),
    transforms.RandomRotation(20),
    transforms.ToTensor(),
    transforms.Normalize([0.5]*3, [0.5]*3)
])

# Dataset
train_data = datasets.ImageFolder(DATASET_DIR, transform=transform)
class_names = train_data.classes
train_loader = DataLoader(train_data, batch_size=BATCH_SIZE, shuffle=True)

# Model: MobileNetV2
model = models.mobilenet_v2(weights=models.MobileNet_V2_Weights.DEFAULT)
model.classifier[1] = nn.Linear(model.classifier[1].in_features, len(class_names))
model = model.to(DEVICE)

# Loss & Optimizer
criterion = nn.CrossEntropyLoss()
optimizer = torch.optim.Adam(model.parameters(), lr=0.001)

# Training loop
# print(f"🔧 Training on {DEVICE} for {EPOCHS} epochs...")
for epoch in range(EPOCHS):
    model.train()
    running_loss = 0.0
    correct, total = 0, 0

    for inputs, labels in tqdm(train_loader, desc=f"Epoch {epoch+1}/{EPOCHS}"):
        inputs, labels = inputs.to(DEVICE), labels.to(DEVICE)

        optimizer.zero_grad()
        outputs = model(inputs)
        loss = criterion(outputs, labels)
        loss.backward()
        optimizer.step()

        running_loss += loss.item()
        _, preds = torch.max(outputs, 1)
        correct += (preds == labels).sum().item()
        total += labels.size(0)

    acc = 100 * correct / total
    # print(f"✅ Epoch {epoch+1}: Loss={running_loss:.4f} | Accuracy={acc:.2f}%")

#Save model and class names
torch.save(model.state_dict(), os.path.join(MODEL_DIR, "model.pt"))
with open(os.path.join(MODEL_DIR, "class_names.txt"), "w") as f:
    f.write("\n".join(class_names))

print("Training complete. Model and classes saved to 'model/'.")
