ai-animal-classification/requirements.txt
flask
torch
torchvision
tensorflow
pillow
numpy
psycopg2-binary
python-dotenv

